Voyagers worker refactor
Refactors voyagers worker into several parts, which should make it easier to add, replace, and maintain the various core elements of USS John Jay's voyage calc.

Rest assured, there's a good reason why I wanted to try this, but we'll see if anything worthwhile comes of it.
